kernel:[vmebus] register devices from vmebus
The vmebus allows to register devices directly by writing a registration
string in a sysfs bus attribute.

There is no need for the VME device driver to register itself the
devices by asking the user to provide module parameters.

Now, all the parameters a given directly to the vmebus.

Also the fw_name module parameter has been removed. The user can use the
char device to re-program the FPGA.

NOTE
* I thought a bit about the fw_name. Users may want to load their
  bitstream immediately before anything else. Is it a real use case?

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>